Minor bug fixes
Abort targetClosestShip when there are no nearby ships (including the player).
Windows doesn't seem to support ctime_r, using ctime instead.
includes.h must always be before common.h
V_SQRT2 should be cast to a float before use with other floats.

+++ Source/Utilities/quadtree.h
@@ -72,7 +72,7 @@ inline bool QuadTree::PossiblyNear(Coordinate point, float distance) {
 	// If the distance to the point is greater than the max range,
 	//   then no collisions are possible
 	// Math should be done in square-space to save time.
-	const float maxrange = V_SQRT2*radius + distance;
+	const float maxrange = static_cast<float>(V_SQRT2)*radius + distance;
 	return( (point-center).GetMagnitudeSquared() <= maxrange*maxrange );
 }
